Split the TC_TLSCERT.py file into separate test files
